Move prometheusreceiver from collector core to collector contrib
