+++ receiver/prometheusreceiver/config_test.go
@@ -25,7 +25,6 @@ import (
 	"github.com/prometheus/prometheus/discovery/kubernetes"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
 	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/config/configtest"
