+++ receiver/prometheusreceiver/internal/metrics_adjuster.go
