+++ receiver/prometheusreceiver/internal/metrics_adjuster_test.go
