+++ receiver/prometheusreceiver/internal/ocastore.go
@@ -23,10 +23,9 @@ import (
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/scrape"
 	"github.com/prometheus/prometheus/storage"
-	"go.uber.org/zap"
-
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/consumer"
+	"go.uber.org/zap"
 )
 
 const (
