+++ receiver/prometheusreceiver/internal/ocastore_test.go
@@ -22,7 +22,6 @@ import (
 	"github.com/prometheus/prometheus/scrape"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
 	"go.opentelemetry.io/collector/config"
 )
