+++ receiver/prometheusreceiver/internal/otlp_metricfamily_test.go
@@ -23,9 +23,8 @@ import (
 	"github.com/prometheus/prometheus/scrape"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"go.uber.org/zap"
-
 	"go.opentelemetry.io/collector/model/pdata"
+	"go.uber.org/zap"
 )
 
 type byLookupMetadataCache map[string]scrape.MetricMetadata
