+++ receiver/prometheusreceiver/internal/otlp_metricsbuilder.go
@@ -24,9 +24,8 @@ import (
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/pkg/textparse"
 	"github.com/prometheus/prometheus/pkg/value"
-	"go.uber.org/zap"
-
 	"go.opentelemetry.io/collector/model/pdata"
+	"go.uber.org/zap"
 )
 
 func isUsefulLabelPdata(mType pdata.MetricDataType, labelKey string) bool {
