+++ receiver/prometheusreceiver/internal/otlp_metricsbuilder_test.go
@@ -23,7 +23,6 @@ import (
 	"github.com/prometheus/prometheus/pkg/textparse"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
 	"go.opentelemetry.io/collector/model/pdata"
 )
