+++ receiver/prometheusreceiver/testdata/config.yaml
