+++ receiver/prometheusreceiver/testdata/config_env.yaml
