+++ receiver/prometheusreceiver/testdata/config_k8s.yaml
