+++ receiver/prometheusreceiver/testdata/config_sd.yaml
