+++ receiver/prometheusreceiver/testdata/invalid-config-prometheus-relabel.yaml
