+++ receiver/prometheusreceiver/testdata/invalid-config-prometheus-section.yaml
