+++ receiver/prometheusreceiver/testdata/invalid-config-prometheus-unsupported-features.yaml
