+++ receiver/prometheusreceiver/testdata/invalid-config-section.yaml
